Modified homepage hero unit to use css grid.
I had a crazy idea to use css grid to get a semi-print styled layout
like a magazine. Right now it's a bit of a mess and I likely need to
restructure the homepage folder. However, it looks much more creative
and I like it. It's **my** website after all.